comments in posts and profile with user posts

+++ controllers/postsController.js
@@ -13,9 +13,12 @@ router.get("/home/posts", adminAuth, (req, res) => {
         include: { model: ShareComment }
      }).then(post => {
         const mappedPosts = post.map(item => {
-            console.log(item)
             return {
-                ...item, 
+                id: item.id,
+                username: item.username,
+                description: item.description,
+                share_comments: item.share_comments,
+                updatedAt: item.updatedAt, 
                 isFromUser: id === item.userId
             }
         })
@@ -27,8 +30,6 @@ router.post("/post/save", adminAuth, (req, res) => {
 
     var description = req.body.description;
     var file = req.body.file;
-/*     console.log(req.body);
-    console.log(req.session.user); */
 
     Post.create({
         description: description,
@@ -45,23 +46,20 @@ router.get("/post/edit/:id", adminAuth, (req, res) => {
     var id = req.params.id;
     Post.findByPk(id).then(post => {
         if(post != undefined) {
-            User.findAll().then(users => {
-                res.render("admin/posts/edit");
-            });
+                res.render("admin/posts/edit", {post: post});
         } else {
-            res.redirect("/home/post")
+            res.redirect("/home/post");
         }
     }).catch(err => {
-        res.redirect("/home/posts")
+        res.redirect("/home/posts");
     });
 });
 
 router.post("/post/update", adminAuth, (req, res) => {
     var id = req.body.id;
     var description = req.body.description;
-    var file = req.body.file;
-
-    Post.update({description: description, file: file}, {
+    console.log(description, id);
+    Post.update({description: description}, {
         where: {
             id: id
         }
@@ -72,27 +70,25 @@ router.post("/post/update", adminAuth, (req, res) => {
     });
 });
 
-router.post("/post/delete", adminAuth, (req, res) => {
+router.post("/post/delete", adminAuth, async (req, res) => {
     var id = req.body.id;
-    if(id != undefined) {
-        if(!isNaN(id)) {
-            Post.destroy({
+    if(id != undefined && !isNaN(id)) {
+            await ShareComment.destroy({
+                where: {
+                    postId: id
+                }
+            }) 
+            await Post.destroy({
                 where: {
                     id: id
                 }
-            }).then (() => {
-                res.redirect("/home/posts");
-            });
-        }else {
-            res.redirect("/home/posts");
+            })                 
         }
-    }else {
-        res.redirect("/home/posts");
-    }
+    res.redirect("/home/posts");
 });
 
 router.get("/profile", (req, res) => {
-    console.log(req.session.user)
+
     Post.findAll({ 
         include: [
            { model: User },
